:tractor: move package model maps
