+++ lib/src/screens/maps_place.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:permission_handler/permission_handler.dart';
 import '../blocs/maps_provider.dart';
-import '../models/list_item_maps_pin_model.dart';
+import '../models/maps/list_item_maps_pin_model.dart';
 import '../data/constants.dart';
 
 class MapsPlace extends StatelessWidget {
